Updated ERD to better reflect business use
Updated relation between plants and items from 'one mandatory to one optional' to 'one mandatory to many optional'. Every plant must still be associated with an item but now plants can be associated with multiple items instead of only one. This alteration allows the case that two plants of the same variety may have different prices or that adult plants and seeds of the same species will be disparate items.